requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
greenlet>=3.0.0
faster-whisper>=1.0.0; python_version < "3.13"
